Fixed a bug which caused optional foreign keys to throw an error.
lib/Instance.js uses 'typeof' to check if a variable is an object.
However, 'typeof null == "object"'. Therefore when opts.data[asc.name] == null
the condition on line 549 would evaluate to true and line 550 will throw an error.
This situation occurs at least when an optional FK is defined (postgres)
and it's value is null. Other situations might trigger this error as well.

This bugfix changes line 549 such that it uses 'instanceof' instead of 'typeof'.