Preserve column names when association_key is a function
Node-orm2 is case-sensitive. Column names are properly escaped so we
can use camelCase columns if we would like to.

There is however one exception - association_keys are automatically
lowercased. This is a bit surprising.

With version 2.1 we go the ability to use functions as
association_keys, but the column names supplied to the function is
already lowercased.

I believe this is wrong - if you are supplying a function as
association_key you obviously want to design your own key format, then
you should be given the original column names and lowercase them
yourself if you want to.

This PR simply removes `.toLowerCase()` if association_key is a
function. That way, we can setup association_keys with proper
camelCase, e.g. someTableId