Set internal default property value to undefined instead of null.
This will prevent null values being explicitly sent to the database when no value was assigned and instead result in the database setting the column to null, or generating a default value.
Properties with an internal value of undefined will still return null when accessed.
Setting a previously filled property to undefined will still set it to null in the DB.
No ORM tests were broken by this change, and as such, the impact of this should be limited ot a very small number of corner cases.

Add PostgreSQL uuid column support.
Allow specifying defaultExpression (eg. uuid_generate_v4() for PostgreSQL or uuid() for MySQL) to be executed by the database engine for generating default values.